blob: split the openbucket example
